fix: use NextJS link component in breadcrumb to prevent page reloads
so whenever trying to navigate from one page to another using breadcurmb links, the app was reloading. due to this, even if datasets was present in memory, the app was reloading resetting everything.

the issue was in breadcrum we were using simple a href tags. instead, we now use nextjs Link component and it's working fine now